# docs/swagger.yaml
basePath: /
definitions:
  calendar.scheduleRequest:
    properties:
      order_id:
        type: string
      start_iso:
        type: string
    type: object
  calendar.slotOption:
    properties:
      label:
        description: 'ex: "06/10 13:00", pronto pra exibir no bot'
        type: string
      start_iso:
        type: string
    type: object
  payment.checkoutRequest:
    properties:
      service_slug:
        type: string
      whatsapp_number:
        type: string
    type: object
  payment.checkoutResponse:
    properties:
      checkout_url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'API de negócio: pagamentos, pedidos e consultoria'
  title: Guerreiro da Fronteira - API
  version: "1.0"
paths:
  /checkout:
    post:
      consumes:
      - application/json
      description: Recebe o número de WhatsApp e o serviço escolhido, cria o pedido
        e retorna a URL de pagamento da Stripe
      parameters:
      - description: Dados do checkout
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.checkoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.checkoutResponse'
        "400":
          description: requisição inválida
          schema:
            type: string
        "500":
          description: erro interno
          schema:
            type: string
      summary: Cria checkout de pagamento
      tags:
      - payment
  /consultoria/agendar:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do agendamento
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/calendar.scheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Confirma o agendamento da consultoria
      tags:
      - consultation
  /consultoria/horarios:
    get:
      parameters:
      - description: ID do pedido
        in: query
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/calendar.slotOption'
              type: array
            type: object
      summary: Lista horários livres para consultoria
      tags:
      - consultation
  /webhooks/stripe:
    post:
      consumes:
      - application/json
      description: 'Recebe eventos da Stripe (ex: checkout.session.completed) e atualiza
        o pedido correspondente'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: assinatura inválida
          schema:
            type: string
        "500":
          description: erro interno
          schema:
            type: string
      summary: Webhook da Stripe
      tags:
      - payment
swagger: "2.0"